Turn caps plugin into folder.
Move more test files to plugins
