+++ src/headless/plugins/roster/tests/presence.js
@@ -2,115 +2,6 @@
 
 // See: https://xmpp.org/rfcs/rfc3921.html
 
-const original_timeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
-
-describe("A sent presence stanza", function () {
-
-    beforeEach(() => (jasmine.DEFAULT_TIMEOUT_INTERVAL = 7000));
-    afterEach(() => (jasmine.DEFAULT_TIMEOUT_INTERVAL = original_timeout));
-
-    it("includes a entity capabilities node",
-            mock.initConverse([], {}, async (done, _converse) => {
-
-        await mock.waitForRoster(_converse, 'current', 0);
-        _converse.api.disco.own.identities.clear();
-        _converse.api.disco.own.features.clear();
-
-        _converse.api.disco.own.identities.add("client", "pc", "Exodus 0.9.1");
-        _converse.api.disco.own.features.add("http://jabber.org/protocol/caps");
-        _converse.api.disco.own.features.add("http://jabber.org/protocol/disco#info");
-        _converse.api.disco.own.features.add("http://jabber.org/protocol/disco#items");
-        _converse.api.disco.own.features.add("http://jabber.org/protocol/muc");
-
-        const presence = _converse.xmppstatus.constructPresence();
-        expect(presence.toLocaleString()).toBe(
-            `<presence xmlns="jabber:client">`+
-                `<priority>0</priority>`+
-                `<c hash="sha-1" node="https://conversejs.org" ver="QgayPKawpkPSDYmwT/WM94uAlu0=" xmlns="http://jabber.org/protocol/caps"/>`+
-            `</presence>`)
-        done();
-    }));
-
-    it("has a given priority", mock.initConverse(['statusInitialized'], {}, (done, _converse) => {
-        const { api } = _converse;
-        let pres = _converse.xmppstatus.constructPresence('online', null, 'Hello world');
-        expect(pres.toLocaleString()).toBe(
-            `<presence xmlns="jabber:client">`+
-                `<status>Hello world</status>`+
-                `<priority>0</priority>`+
-                `<c hash="sha-1" node="https://conversejs.org" ver="PxXfr6uz8ClMWIga0OB/MhKNH/M=" xmlns="http://jabber.org/protocol/caps"/>`+
-            `</presence>`
-        );
-        api.settings.set('priority', 2);
-        pres = _converse.xmppstatus.constructPresence('away', null, 'Going jogging');
-        expect(pres.toLocaleString()).toBe(
-            `<presence xmlns="jabber:client">`+
-                `<show>away</show>`+
-                `<status>Going jogging</status>`+
-                `<priority>2</priority>`+
-                `<c hash="sha-1" node="https://conversejs.org" ver="PxXfr6uz8ClMWIga0OB/MhKNH/M=" xmlns="http://jabber.org/protocol/caps"/>`+
-            `</presence>`
-        );
-
-        api.settings.set('priority', undefined);
-        pres = _converse.xmppstatus.constructPresence('dnd', null, 'Doing taxes');
-        expect(pres.toLocaleString()).toBe(
-            `<presence xmlns="jabber:client">`+
-                `<show>dnd</show>`+
-                `<status>Doing taxes</status>`+
-                `<priority>0</priority>`+
-                `<c hash="sha-1" node="https://conversejs.org" ver="PxXfr6uz8ClMWIga0OB/MhKNH/M=" xmlns="http://jabber.org/protocol/caps"/>`+
-            `</presence>`
-        );
-        done();
-    }));
-
-    it("includes the saved status message",
-        mock.initConverse([], {}, async (done, _converse) => {
-
-        const { u, Strophe } = converse.env;
-        mock.openControlBox(_converse);
-        spyOn(_converse.connection, 'send').and.callThrough();
-
-        const cbview = _converse.chatboxviews.get('controlbox');
-        const change_status_el = await u.waitUntil(() => cbview.querySelector('.change-status'));
-        change_status_el.click()
-        let modal = _converse.api.modal.get('modal-status-change');
-        await u.waitUntil(() => u.isVisible(modal.el), 1000);
-        const msg = 'My custom status';
-        modal.el.querySelector('input[name="status_message"]').value = msg;
-        modal.el.querySelector('[type="submit"]').click();
-
-        const sent_stanzas = _converse.connection.sent_stanzas;
-        let sent_presence = await u.waitUntil(() => sent_stanzas.filter(s => Strophe.serialize(s).match('presence')).pop());
-        expect(Strophe.serialize(sent_presence))
-            .toBe(`<presence xmlns="jabber:client">`+
-                    `<status>My custom status</status>`+
-                    `<priority>0</priority>`+
-                    `<c hash="sha-1" node="https://conversejs.org" ver="PxXfr6uz8ClMWIga0OB/MhKNH/M=" xmlns="http://jabber.org/protocol/caps"/>`+
-                    `</presence>`)
-        await u.waitUntil(() => modal.el.getAttribute('aria-hidden') === "true");
-        await u.waitUntil(() => !u.isVisible(modal.el));
-
-        cbview.querySelector('.change-status').click()
-        modal = _converse.api.modal.get('modal-status-change');
-        await u.waitUntil(() => modal.el.getAttribute('aria-hidden') === "false", 1000);
-        modal.el.querySelector('label[for="radio-busy"]').click(); // Change status to "dnd"
-        modal.el.querySelector('[type="submit"]').click();
-        await u.waitUntil(() => sent_stanzas.filter(s => Strophe.serialize(s).match('presence')).length === 2);
-        sent_presence = sent_stanzas.filter(s => Strophe.serialize(s).match('presence')).pop();
-        expect(Strophe.serialize(sent_presence))
-            .toBe(
-                `<presence xmlns="jabber:client">`+
-                    `<show>dnd</show>`+
-                    `<status>My custom status</status>`+
-                    `<priority>0</priority>`+
-                    `<c hash="sha-1" node="https://conversejs.org" ver="PxXfr6uz8ClMWIga0OB/MhKNH/M=" xmlns="http://jabber.org/protocol/caps"/>`+
-                `</presence>`)
-        done();
-    }));
-});
-
 describe("A received presence stanza", function () {
 
     it("has its priority taken into account",
